+++ defaults_js.go
@@ -1,9 +1,7 @@
-//go:build !js
-// +build !js
+//go:build js
+// +build js
 
 package libp2p
-
-// This file contains all the default configuration options.
 
 import (
 	"crypto/rand"
@@ -15,52 +13,22 @@ import (
 	"github.com/libp2p/go-libp2p/p2p/net/connmgr"
 	"github.com/libp2p/go-libp2p/p2p/net/swarm"
 	"github.com/libp2p/go-libp2p/p2p/security/noise"
-	tls "github.com/libp2p/go-libp2p/p2p/security/tls"
-	quic "github.com/libp2p/go-libp2p/p2p/transport/quic"
-	"github.com/libp2p/go-libp2p/p2p/transport/tcp"
-	libp2pwebrtc "github.com/libp2p/go-libp2p/p2p/transport/webrtc"
-	ws "github.com/libp2p/go-libp2p/p2p/transport/websocket"
-	webtransport "github.com/libp2p/go-libp2p/p2p/transport/webtransport"
 	"github.com/prometheus/client_golang/prometheus"
-
-	"github.com/multiformats/go-multiaddr"
 )
 
-// DefaultSecurity is the default security option.
-//
-// Useful when you want to extend, but not replace, the supported transport
-// security protocols.
 var DefaultSecurity = ChainOptions(
-	Security(tls.ID, tls.New),
 	Security(noise.ID, noise.New),
 )
 
-// DefaultMuxers configures libp2p to use the stream connection multiplexers.
-//
-// Use this option when you want to *extend* the set of multiplexers used by
-// libp2p instead of replacing them.
 var DefaultMuxers = Muxer(yamux.ID, yamux.DefaultTransport)
 
-// DefaultTransports are the default libp2p transports.
-//
-// Use this option when you want to *extend* the set of transports used by
-// libp2p instead of replacing them.
-var DefaultTransports = ChainOptions(
-	Transport(tcp.NewTCPTransport),
-	Transport(quic.NewTransport),
-	Transport(ws.New),
-	Transport(webtransport.New),
-	Transport(libp2pwebrtc.New),
-)
+var DefaultTransports = ChainOptions(NoTransports)
 
 // DefaultPrivateTransports are the default libp2p transports when a PSK is supplied.
 //
 // Use this option when you want to *extend* the set of transports used by
 // libp2p instead of replacing them.
-var DefaultPrivateTransports = ChainOptions(
-	Transport(tcp.NewTCPTransport),
-	Transport(ws.New),
-)
+var DefaultPrivateTransports = ChainOptions(NoTransports)
 
 // DefaultPeerstore configures libp2p to use the default peerstore.
 var DefaultPeerstore Option = func(cfg *Config) error {
@@ -81,27 +49,7 @@ var RandomIdentity = func(cfg *Config) error {
 }
 
 // DefaultListenAddrs configures libp2p to use default listen address.
-var DefaultListenAddrs = func(cfg *Config) error {
-	addrs := []string{
-		"/ip4/0.0.0.0/tcp/0",
-		"/ip4/0.0.0.0/udp/0/quic-v1",
-		"/ip4/0.0.0.0/udp/0/quic-v1/webtransport",
-		"/ip4/0.0.0.0/udp/0/webrtc-direct",
-		"/ip6/::/tcp/0",
-		"/ip6/::/udp/0/quic-v1",
-		"/ip6/::/udp/0/quic-v1/webtransport",
-		"/ip6/::/udp/0/webrtc-direct",
-	}
-	listenAddrs := make([]multiaddr.Multiaddr, 0, len(addrs))
-	for _, s := range addrs {
-		addr, err := multiaddr.NewMultiaddr(s)
-		if err != nil {
-			return err
-		}
-		listenAddrs = append(listenAddrs, addr)
-	}
-	return cfg.Apply(ListenAddrs(listenAddrs...))
-}
+var DefaultListenAddrs = NoListenAddrs
 
 // DefaultEnableRelay enables relay dialing and listening by default.
 var DefaultEnableRelay = func(cfg *Config) error {
